Change underlying type for POINT from float to double
Also, use a different representation for null---not (NaN, NaN), but
(360.0, 360.0).

Made the Java client return a null object reference instead of returning an
object with the lat and long set to magic numbers.

Added POINT and GEOGRAPHY to the round trip DDL tests.

+++ src/frontend/org/voltdb/types/GeographyValue.java
@@ -245,7 +245,7 @@ public class GeographyValue {
                         throw new IllegalArgumentException(msgPrefix + "missing longitude in lat long pair");
                     }
                     double lng = tokenizer.nval;
-                    currentLoop.add(new PointType((float)lat, (float)lng));
+                    currentLoop.add(new PointType(lat, lng));
 
                     token = tokenizer.nextToken();
                     if (token != ',') {
